Fixed dependency issues that were caused by creating a custom user model

+++ stocks/models.py
@@ -14,7 +14,7 @@ class Portfolio(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     stock_ticker = models.CharField(max_length=20)
     num_of_shares = models.IntegerField()
-    price_bought = models.DecimalField(max_digits=10, decimal_places=2)
+    price_bought = models.FloatField()
     date_bought = models.DateTimeField(default=timezone.now)
 
 #database table to store a users transaction information
